backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
psycopg2-binary==2.9.9
pandas==2.1.4
prophet==1.1.5
cmdstanpy==1.2.0
python-dotenv==1.0.0
SQLAlchemy==2.0.23
numpy<2.0.0
